requirements.txt
streamlit
google-api-python-client

llama-index-core

llama-index-readers-google

llama-index-embeddings-huggingface

llama-index-llms-openai
